feat: extract out db
